pyproject.toml
[tool.poetry]
name = "ice-cream-dataops"
version = "0.1.0"
description = ""
authors = ["Ferrao"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
cognite-toolkit = "0.3.0b4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
